Bug fixed and should be how I like it.  Only thing I know of left is to make it device specific

+++ src/main/kotlin/com/davidcorrado/kotlinrest/controller/TaskController.kt
@@ -3,26 +3,33 @@ package com.davidcorrado.kotlinrest.controller
 import com.davidcorrado.kotlinrest.data.Task
 import com.davidcorrado.kotlinrest.data.TaskRepository
 import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.data.crossstore.ChangeSetPersister
+import org.springframework.http.HttpStatus
 import org.springframework.web.bind.annotation.*
+import org.springframework.web.client.HttpClientErrorException
 
 @RestController
-@RequestMapping( "/api/v1/task")
+@RequestMapping( "/api/v1/tasks")
 class TaskController {
 
     @Autowired lateinit var taskRepo: TaskRepository
 
     @GetMapping()
-    fun getAll(): Any {
+    fun getAll(): Iterable<Task> {
         return taskRepo.findAll()
     }
 
     @GetMapping("/{id}")
-    fun getByID(@PathVariable id: Long) {
-        taskRepo.findOne(id)
+    fun getByID(@PathVariable id: Long): Task {
+        if(taskRepo.exists(id)) {
+            return taskRepo.findOne(id)
+        } else {
+            throw HttpClientErrorException(HttpStatus.NOT_FOUND,"Task not found.")
+        }
     }
 
     @PostMapping()
-    fun save(@RequestParam task: Task): Any {
+    fun save(@RequestBody task: Task): Task {
         return taskRepo.save(task)
     }
 
@@ -33,6 +40,8 @@ class TaskController {
 
     @DeleteMapping("/{id}")
     fun deleteByID(@PathVariable id: Long) {
-        taskRepo.delete(id)
+        if(taskRepo.exists(id)) {
+            taskRepo.delete(id)
+        }
     }
 }
